pyproject.toml
[project]
name = "airbnb-features"
version = "0.1.0"
description = "Feature Store para dados Airbnb no Databricks"
requires-python = ">=3.12"
dependencies = [
    "databricks-connect>=18.0,<18.1",
    "databricks-sdk",
    "databricks-feature-engineering",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "ipykernel",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airbnb_features"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
